pyproject.toml
[project]
name = "devcontext"
version = "0.1.0"
description = "Resume any project in 30 seconds. Open source CLI for developer context management."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Bitfuturistic Solutions", email = "hello@bitfuturistic.com" }
]
keywords = ["developer", "productivity", "context", "cli", "project-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "gitpython>=3.1.0",
    "ollama>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
devctx = "devcontext.cli:main"

[project.urls]
Homepage = "https://github.com/tsunderland/devcontext"
Documentation = "https://github.com/tsunderland/devcontext#readme"
Repository = "https://github.com/tsunderland/devcontext"
Issues = "https://github.com/tsunderland/devcontext/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/devcontext"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
